Day283✅/requirements.txt
django 
djangorestframework 
pandas 
scikit-learn 
joblib
numpy